Separating Boolean logical instructions into separate files.
